RAT-FUD/requirement.txt
opencv-python==4.7.0.68
pynput==1.7.6
pyautogui==0.9.53
pillow==9.4.0
numpy==1.24.2
pygame--pre
